rename: view 폴더 하위로 recipe 폴더 이동
