+++ app/src/main/res/layout/fragment_recipemain.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".community.CommunityFragment"
+    tools:context=".view.community.CommunityFragment"
     android:orientation="vertical"
 
     >
