+++ app/src/main/res/layout/fragment_scrap.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".recipe.ScrapFragment">
+    tools:context=".view.recipe.ScrapFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/my_toolbar"
